Refactor parse management command
